refactor: move logger_setup to utils/, move legacy-only files to legacy/
