+++ simba/core/chemistry/mces/mces_computation.py
@@ -15,8 +15,8 @@ from simba.core.chemistry.mces_loader.load_mces import LoadMCES
 from simba.core.data.molecular_pairs import MolecularPairsSet
 from simba.core.data.molecule_pairs_opt import MoleculePairsOpt
 from simba.core.data.spectrum import SpectrumExt
-from simba.logger_setup import logger
-from simba.train_utils import TrainUtils
+from simba.core.training.train_utils import TrainUtils
+from simba.utils.logger_setup import logger
 
 
 class MCES:
